added folder under pages to hold all CreateJobPosting pages, renamed Create styling sheet
